[styled-components] Move React Native-dependent types to separate package
