+++ types/styled-components/ts3.6/tsconfig.json
@@ -4,6 +4,7 @@
         "forceConsistentCasingInFileNames": true,
         "jsx": "react",
         "lib": [
+            "dom",
             "es6"
         ],
         "module": "commonjs",
@@ -20,7 +21,6 @@
     "files": [
         "index.d.ts",
         "test/index.tsx",
-        "test/native.tsx",
         "test/macro.tsx"
     ]
 }
